(refactor): move chart to generic component
