--- frontend/src/app/reports/components/derivations-report/derivations-report.component.scss
@@ -1,7 +1,0 @@
-.title {
-    color: #4A4F52;
-}
-
-.btn-purple {
-    background-color: #8C008C !important;
-}
