+++ frontend/src/app/reports/components/report/report.component.html
@@ -1,0 +1,42 @@
+<div class="col-md-8 mx-auto mt-5">
+    <h1 class="text-center title">{{title}}</h1>
+    <div *ngIf="errorMessage" class="alert alert-danger text-center" role="alert">
+        <b>{{errorMessage}}</b>
+    </div>
+    <form [formGroup]="formGroup">
+        <div class="form-row justify-content-center">
+            <div class="form-group col-md-3">
+                <label for="fromDate">Fecha desde</label>
+                <input type="date" class="form-control" formControlName="fromDate"
+                    [ngClass]="{ 'is-invalid': submitted && formGroup.controls.fromDate.errors }"
+                    (change)="dateChange()">
+
+                <div *ngIf="formGroup.controls.fromDate.errors?.futureDate" class="invalid-feedback">
+                    La fecha no puede ser futura
+                </div>
+                <div class="invalid-feedback" *ngIf="formGroup.controls.fromDate.errors?.required">
+                    Debe ingresar una fecha
+                </div>
+            </div>
+            <div class="form-group col-md-3">
+                <label for="toDate">Fecha hasta</label>
+                <input type="date" class="form-control" formControlName="toDate"
+                    [ngClass]="{ 'is-invalid': submitted && formGroup.controls.toDate.errors }" (change)="dateChange()">
+                <div *ngIf="formGroup.controls.toDate.errors?.futureDate" class="invalid-feedback">
+                    La fecha no puede ser futura
+                </div>
+                <div class="invalid-feedback" *ngIf="formGroup.controls.toDate.errors?.invalidToDate">
+                    La fecha hasta debe ser posterior a la fecha desde
+                </div>
+                <div class="invalid-feedback" *ngIf="formGroup.controls.toDate.errors?.required">
+                    Debe ingresar una fecha
+                </div>
+            </div>
+        </div>
+    </form>
+    <div class="chart-wrapper" *ngIf="barChartLabels.length > 0">
+        <canvas baseChart [datasets]="barChartData" [labels]="barChartLabels" [options]="barChartOptions"
+            [plugins]="barChartPlugins" [legend]="barChartLegend" [chartType]="barChartType">
+        </canvas>
+    </div>
+</div>
