+++ frontend/src/app/reports/components/report/report.component.scss
@@ -1,0 +1,3 @@
+.title {
+    color: #4A4F52;
+}
